avif backend records the original source bytes, not the letterboxed pngs
the avif path letterboxes every image to a png in a tempdir and handed
only those pngs to encode_avif, which summed them as source_bytes. on the
38627-card jpeg corpus that recorded 21450566470 bytes and a 18.61x ratio
for a 3975063106-byte source whose true ratio is 3.45x. the backend now
passes the original file sum in, the encoder keeps the png sum apart as
letterboxed_input_bytes, and compression_ratio is over the originals for
every backend.

both touched modules were over the line limit, so each is split along a
single responsibility: image_encode_still.py (avif + jxl encoders),
image_gop_probe.py (the gop probe), image_backends_av1.py (the stream
backend). imports, ruff targets and tests updated.
