+++ tests/test_quality_gate.py
@@ -20,7 +20,7 @@ sys.path.insert(0, str(REPO / "python"))
 
 import numpy as np
 from forge.build_spec import MediaSpec, QualitySpec
-from forge.image_encode import encode_jxl_dir
+from forge.image_encode_still import encode_jxl_dir
 
 HAVE = all(shutil.which(t) for t in ("ffmpeg", "ssimulacra2", "cjxl", "djxl"))
